feat(admin): make /configs/validate available in all modes
Extract batch configuration validation logic from standalone.lua into
a new shared module (admin/config_validate.lua) and register the
POST /apisix/admin/configs/validate endpoint for etcd mode as well.

Previously this endpoint was only available in standalone (yaml) mode.
Clients using etcd mode (e.g., ADC, Ingress Controller) had no way to
perform batch validation that includes plugin check_schema() advanced
checks (cross-field constraints, regex compilation, expression parsing,
recursive sub-plugin validation, etc.).

Changes:
- New module: apisix/admin/config_validate.lua
  Owns validate_configuration(), check_conf(), check_duplicate(), and
  the validate() HTTP handler. Shared by both standalone and etcd modes.
- standalone.lua: delegates to config_validate module, removing ~140
  lines of duplicated logic while keeping standalone-specific code
  (get/update/head, patch_schema, config sync).
- init.lua: registers POST /apisix/admin/configs/validate in the etcd
  mode route table, placed before the wildcard route for correct
  matching priority.
